Moved utility functions for commands into seperate file
